updated to use int

+++ Circumference/main.py
@@ -1,5 +1,5 @@
 pi = 3.141592653589793
-radius = float(input("Input radius:"))
+radius = int(input("Input radius:"))
 circumference = pi * radius * 2
 area = pi * radius * radius
 print(f"The circumference is: {circumference}")
